archive/tekup-gmail-automation-archived-2025-10-22/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tekup-gmail-automation"
version = "1.2.0"
description = "TekUp Gmail Automation - Intelligent PDF forwarding and receipt processing"
authors = [
    {name = "TekUp", email = "info@tekup.dk"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=0.5.0",
    "google-auth>=2.0.0",
    "requests>=2.28.0",
    "python-dotenv>=0.19.0",
    "pydantic>=1.10.0",
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-multipart>=0.0.5",
    "aiofiles>=0.8.0",
    "pillow>=9.0.0",
    "pytesseract>=0.3.8",
    "opencv-python>=4.6.0",
    "numpy>=1.21.0",
    "pandas>=1.4.0",
    "openpyxl>=3.0.0",
    "schedule>=1.2.0",
    "python-dateutil>=2.8.0",
    "pytz>=2022.1",
    "email-validator>=1.2.0",
    "jinja2>=3.1.0",
    "markdown>=3.4.0",
    "pyyaml>=6.0",
    "toml>=0.10.2",
    "click>=8.1.0",
    "rich>=12.5.0",
    "loguru>=0.6.0",
    "tenacity>=8.0.0",
    "httpx>=0.23.0",
    "aiohttp>=3.8.0",
    "websockets>=10.0",
    "redis>=4.3.0",
    "celery>=5.2.0",
    "sqlalchemy>=1.4.0",
    "alembic>=1.8.0",
    "psycopg2-binary>=2.9.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.19.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pre-commit>=2.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.19.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pre-commit>=2.17.0",
]

[project.urls]
Homepage = "https://github.com/tekup/tekup-gmail-automation"
Documentation = "https://github.com/tekup/tekup-gmail-automation/blob/main/README.md"
Repository = "https://github.com/tekup/tekup-gmail-automation"
Issues = "https://github.com/tekup/tekup-gmail-automation/issues"

[project.scripts]
tekup-gmail = "src.core.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
